requirements.txt
# TensorFlow & Keras (macOS-compatible)
tensorflow-macos==2.11.0
keras==2.11.0
protobuf==3.19.6


# LangChain, LLM, Embeddings (Together API — no transformers needed)
langchain==0.1.16
langchain_community==0.0.34
faiss-cpu==1.7.4
sentence-transformers==2.2.2
torch==1.13.1


# UI + Utils
streamlit==1.20.0
pandas==1.5.3
scikit-learn==1.2.2
pdfplumber==0.10.3